devtools/server/tests/mochitest/test_inspector-remove.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript" src="inspector-helpers.js"></script>
  <script type="application/javascript">
"use strict";

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
};

let gWalker = null;
let gTarget = null;
let gInspectee = null;

function assertOwnership() {
  return assertOwnershipTrees(gWalker);
}

function ignoreNode(node) {
  // Duplicate the walker logic to skip blank nodes...
  return node.nodeType === Node.TEXT_NODE &&
    !/[^\s]/.test(node.nodeValue);
}

addTest(async function setup() {
  const url = document.getElementById("inspectorContent").href;
  const { target, doc } = await attachURL(url);
  gInspectee = doc;
  const inspector = target.getInspector();
  gWalker = await inspector.getWalker();
  ok(gWalker, "getWalker() should return an actor.");
  gTarget = target;
  runNextTest();
});

addTest(function testRemoveSubtree() {
  let originalOwnershipSize = 0;
  let longlist = null;
  let longlistID = null;

  let nextSibling = gInspectee.querySelector("#longlist").nextSibling;
  while (nextSibling && ignoreNode(nextSibling)) {
    nextSibling = nextSibling.nextSibling;
  }

  let previousSibling = gInspectee.querySelector("#longlist").previousSibling;
  while (previousSibling && ignoreNode(previousSibling)) {
    previousSibling = previousSibling.previousSibling;
  }

  promiseDone(gWalker.querySelector(gWalker.rootNode, "#longlist").then(listFront => {
    longlist = listFront;
    longlistID = longlist.actorID;
  }).then(() => {
    return gWalker.children(longlist);
  }).then((items)=> {
    originalOwnershipSize = assertOwnership();
    // Here is how the ownership tree is summed up:
    // #document                      1
    //   <html>                       1
    //     <body>                     1
    //       <div id=longlist>        1
    //         <div id=a>a</div>   26*2 (each child plus it's singleTextChild)
    //         ...
    //         <div id=z>z</div>
    //                             -----
    //                               56
    is(originalOwnershipSize, 56, "Correct number of items in ownership tree");
    return gWalker.removeNode(longlist);
  }).then(siblings => {
    is(siblings.previousSibling.rawNode(), previousSibling,
       "Should have returned the previous sibling.");
    is(siblings.nextSibling.rawNode(), nextSibling,
       "Should have returned the next sibling.");
    return waitForMutation(gWalker, isChildList);
  }).then(() => {
    // Our ownership size should now be 51 fewer (we forgot about #longlist + 26
    // children + 26 singleTextChild nodes, but learned about #longlist's
    // prev/next sibling)
    const newOwnershipSize = assertOwnership();
    is(newOwnershipSize, originalOwnershipSize - 51,
      "Ownership tree should be lower");
    // Now verify that some nodes have gone away
    return checkMissing(gTarget, longlistID);
  }).then(runNextTest));
});

addTest(function cleanup() {
  gWalker = null;
  gTarget = null;
  gInspectee = null;
  runNextTest();
});
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
